fixed playcontrol, uibackend callout auth

+++ PlayControl.py
@@ -10,21 +10,18 @@ class PlayControl(AlarmService):
         self.className = "Player"
 
     def start(self):
-        #self.pid = 1234
-        # if not playerPid:
-        #print(playerCommand)
+        if self.pid:
+            self.debug("A player instance is running, stopping it first")
+            self.stop()
         self.debug("Start playing music")
         proc = subprocess.Popen(['/usr/bin/nohup','/usr/bin/mpg123','-@','/home/kemkem/playlist/list','-Z','&'])
         self.pid = proc.pid
-        #print(playerPid)
-        # else:
-        #     print('player is already running')
+
 
     def stop(self):
-        # if playerPid:
-        self.debug("Stop playing music")
-        #print(self.pid)
-        os.kill(self.pid, signal.SIGTERM)
-        self.pid = None
-        # else:
-        #     print('player is not running')
+        if self.pid:
+            self.debug("Stop playing music")
+            os.kill(self.pid, signal.SIGTERM)
+            self.pid = None
+        else:
+            self.debug("No player running")
